ElixirOC v0.1.0 ElixirOC.Worker

Worker that does all the requests.

Functions

loop()

Sends the requested information to the coordinator pid.

route_summary(bus_stop)

route_summary(integer) :: {integer, String.t}

Returns a map of the route summary of the requested bus stop.

Examples

iex> ElixirOC.Worker.route_summary(7659)
%{1 => "Ottawa-Rockcliffe", 7 => "St-Laurent"}

route_summary(bus_stop, route_no)

route_summary(integer, integer) :: String.t

Returns a string of the final stop the bus is heading for of the requested bus stop and route number.

Examples

iex> ElixirOC.Worker.route_summary(7659, 1)
"Ottawa-Rockcliffe"